GXDLMSConverter added to show enum values as string. date time DAYLIGHT_SAVE_ACTIVE (summer time) and deviation updates. Push Setup object added. Message Handler object added. MBus master port setup object added.

+++ src/gurux/dlms/objects/GXDLMSIp4Setup.java
@@ -268,7 +268,7 @@ public class GXDLMSIp4Setup extends GXDLMSObject implements IGXDLMSBase
         }
         if (index == 3)
         {
-            return DataType.UINT16;
+            return DataType.UINT32;
         }
         if (index == 4)
         {
